backend/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1

# QuickBooks Online API integration
requests>=2.31.0
python-dotenv>=1.0.0

# FastAPI backend (React frontend)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=21.2.0
python-multipart>=0.0.9

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Cross-process file locking (consolidator, logger, session files)
filelock>=3.14.0
